Created Table: Roster
To hold list of users that are enrolled for an item(course) for a specific offering

SAP Def: Enrollment and attendance in scheduled offering - live learning events led by an instructor - define a relationship between a user and that scheduled offering. For example, the user can be enrolled into the scheduled offering or that enrollment can be canceled. Instructors might want a list of all users enrolled (a course roster) and might want to mark who was enrolled and then attended the event. These are all stories about connecting users to live events or looking for users in live events.